requirements.txt
#MAIN
streamlit==1.28.0
tensorflow-cpu==2.12.0
numpy==1.23.5
pillow==10.1.0
opencv-python-headless==4.8.1.78
torch==2.2.0
ultralytics==8.1.30

#SUB
# streamlit==1.28.0
# tensorflow-cpu==2.20.0
# ultralytics==8.1.2
# Pillow==10.0.0
# numpy==1.26.0
# opencv-python-headless==4.9.0.74
